pyproject.toml
[project]
name = "content-accounts"
version = "1.0.0"
description = "Multi-account content generation framework for social media carousels"
authors = [{name = "Grant Goldman"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "openai>=1.0.0",
    "pillow>=10.0.0",
    "pilmoji>=2.0.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.31.0",
    "questionary>=2.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "mypy>=1.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
content-generate = "cli.generate:main"
content-create = "cli.create_account:main"
content-validate = "cli.validate:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
